+++ sdk/identity/azure-identity/tests/test_shared_cache_credential.py
@@ -4,7 +4,7 @@
 # ------------------------------------
 from azure.core.exceptions import ClientAuthenticationError
 from azure.core.pipeline.policies import SansIOHTTPPolicy
-from azure.identity import KnownAuthorities, SharedTokenCacheCredential
+from azure.identity import CredentialUnavailableError, KnownAuthorities, SharedTokenCacheCredential
 from azure.identity._internal.shared_token_cache import (
     KNOWN_ALIASES,
     MULTIPLE_ACCOUNTS,
@@ -63,13 +63,15 @@ def test_user_agent():
 
 
 def test_empty_cache():
-    with pytest.raises(ClientAuthenticationError, match=NO_ACCOUNTS):
+    """the credential should raise CredentialUnavailableError when the cache is empty"""
+
+    with pytest.raises(CredentialUnavailableError, match=NO_ACCOUNTS):
         SharedTokenCacheCredential(_cache=TokenCache()).get_token("scope")
-    with pytest.raises(ClientAuthenticationError, match=NO_ACCOUNTS):
+    with pytest.raises(CredentialUnavailableError, match=NO_ACCOUNTS):
         SharedTokenCacheCredential(_cache=TokenCache(), username="not@cache").get_token("scope")
-    with pytest.raises(ClientAuthenticationError, match=NO_ACCOUNTS):
+    with pytest.raises(CredentialUnavailableError, match=NO_ACCOUNTS):
         SharedTokenCacheCredential(_cache=TokenCache(), tenant_id="not-cached").get_token("scope")
-    with pytest.raises(ClientAuthenticationError, match=NO_ACCOUNTS):
+    with pytest.raises(CredentialUnavailableError, match=NO_ACCOUNTS):
         SharedTokenCacheCredential(_cache=TokenCache(), tenant_id="not-cached", username="not@cache").get_token("scope")
 
 
@@ -81,7 +83,7 @@ def test_no_matching_account_for_username():
     account = get_account_event(username=upn, uid="uid", utid=tenant, refresh_token="refresh-token")
     cache = populated_cache(account)
 
-    with pytest.raises(ClientAuthenticationError) as ex:
+    with pytest.raises(CredentialUnavailableError) as ex:
         SharedTokenCacheCredential(_cache=cache, username="not" + upn).get_token("scope")
 
     assert ex.value.message.startswith(NO_MATCHING_ACCOUNTS[: NO_MATCHING_ACCOUNTS.index("{")])
@@ -96,7 +98,7 @@ def test_no_matching_account_for_tenant():
     account = get_account_event(username=upn, uid="uid", utid=tenant, refresh_token="refresh-token")
     cache = populated_cache(account)
 
-    with pytest.raises(ClientAuthenticationError) as ex:
+    with pytest.raises(CredentialUnavailableError) as ex:
         SharedTokenCacheCredential(_cache=cache, tenant_id="not-" + tenant).get_token("scope")
 
     assert ex.value.message.startswith(NO_MATCHING_ACCOUNTS[: NO_MATCHING_ACCOUNTS.index("{")])
@@ -111,7 +113,7 @@ def test_no_matching_account_for_tenant_and_username():
     account = get_account_event(username=upn, uid="uid", utid=tenant, refresh_token="refresh-token")
     cache = populated_cache(account)
 
-    with pytest.raises(ClientAuthenticationError) as ex:
+    with pytest.raises(CredentialUnavailableError) as ex:
         SharedTokenCacheCredential(_cache=cache, tenant_id="not-" + tenant, username="not" + upn).get_token("scope")
 
     assert ex.value.message.startswith(NO_MATCHING_ACCOUNTS[: NO_MATCHING_ACCOUNTS.index("{")])
@@ -134,13 +136,13 @@ def test_no_matching_account_for_tenant_or_username():
     transport = Mock(side_effect=Exception())  # credential shouldn't use the network
 
     credential = SharedTokenCacheCredential(username=upn_a, tenant_id=tenant_b, _cache=cache, transport=transport)
-    with pytest.raises(ClientAuthenticationError) as ex:
+    with pytest.raises(CredentialUnavailableError) as ex:
         credential.get_token("scope")
     assert ex.value.message.startswith(NO_MATCHING_ACCOUNTS[: NO_MATCHING_ACCOUNTS.index("{")])
     assert upn_a in ex.value.message and tenant_b in ex.value.message
 
     credential = SharedTokenCacheCredential(username=upn_b, tenant_id=tenant_a, _cache=cache, transport=transport)
-    with pytest.raises(ClientAuthenticationError) as ex:
+    with pytest.raises(CredentialUnavailableError) as ex:
         credential.get_token("scope")
     assert ex.value.message.startswith(NO_MATCHING_ACCOUNTS[: NO_MATCHING_ACCOUNTS.index("{")])
     assert upn_b in ex.value.message and tenant_a in ex.value.message
@@ -232,11 +234,11 @@ def test_no_refresh_token():
     transport = Mock(side_effect=Exception())  # credential shouldn't use the network
 
     credential = SharedTokenCacheCredential(_cache=cache, transport=transport)
-    with pytest.raises(ClientAuthenticationError, match=NO_ACCOUNTS):
+    with pytest.raises(CredentialUnavailableError, match=NO_ACCOUNTS):
         credential.get_token("scope")
 
     credential = SharedTokenCacheCredential(_cache=cache, transport=transport, username="not@cache")
-    with pytest.raises(ClientAuthenticationError, match=NO_ACCOUNTS):
+    with pytest.raises(CredentialUnavailableError, match=NO_ACCOUNTS):
         credential.get_token("scope")
 
 
@@ -252,7 +254,7 @@ def test_two_accounts_no_username_or_tenant():
     # credential can't select an identity => it shouldn't use the network
     transport = Mock(side_effect=Exception())
 
-    # two users in the cache, no username specified -> ClientAuthenticationError
+    # two users in the cache, no username specified -> CredentialUnavailableError
     credential = SharedTokenCacheCredential(_cache=cache, transport=transport)
     with pytest.raises(ClientAuthenticationError, match=MULTIPLE_ACCOUNTS) as ex:
         credential.get_token("scope")
@@ -341,7 +343,7 @@ def test_same_username_different_tenants():
     # with no tenant specified the credential can't select an identity
     transport = Mock(side_effect=Exception())  # (so it shouldn't use the network)
     credential = SharedTokenCacheCredential(username=upn, _cache=cache, transport=transport)
-    with pytest.raises(ClientAuthenticationError) as ex:
+    with pytest.raises(CredentialUnavailableError) as ex:
         credential.get_token("scope")
 
     assert ex.value.message.startswith(MULTIPLE_MATCHING_ACCOUNTS[: MULTIPLE_MATCHING_ACCOUNTS.index("{")])
@@ -384,7 +386,7 @@ def test_same_tenant_different_usernames():
     # with no username specified the credential can't select an identity
     transport = Mock(side_effect=Exception())  # (so it shouldn't use the network)
     credential = SharedTokenCacheCredential(tenant_id=tenant_id, _cache=cache, transport=transport)
-    with pytest.raises(ClientAuthenticationError) as ex:
+    with pytest.raises(CredentialUnavailableError) as ex:
         credential.get_token("scope")
 
     assert ex.value.message.startswith(MULTIPLE_MATCHING_ACCOUNTS[: MULTIPLE_MATCHING_ACCOUNTS.index("{")])