+++ sdk/identity/azure-identity/tests/test_shared_cache_credential_async.py
@@ -6,7 +6,7 @@ from unittest.mock import Mock, patch
 
 from azure.core.exceptions import ClientAuthenticationError
 from azure.core.pipeline.policies import SansIOHTTPPolicy
-from azure.identity import KnownAuthorities
+from azure.identity import CredentialUnavailableError, KnownAuthorities
 from azure.identity.aio import SharedTokenCacheCredential
 from azure.identity._internal.shared_token_cache import (
     KNOWN_ALIASES,
@@ -109,13 +109,15 @@ async def test_user_agent():
 
 @pytest.mark.asyncio
 async def test_empty_cache():
-    with pytest.raises(ClientAuthenticationError, match=NO_ACCOUNTS):
+    """the credential should raise CredentialUnavailableError when the cache is empty"""
+
+    with pytest.raises(CredentialUnavailableError, match=NO_ACCOUNTS):
         await SharedTokenCacheCredential(_cache=TokenCache()).get_token("scope")
-    with pytest.raises(ClientAuthenticationError, match=NO_ACCOUNTS):
+    with pytest.raises(CredentialUnavailableError, match=NO_ACCOUNTS):
         await SharedTokenCacheCredential(_cache=TokenCache(), username="not@cache").get_token("scope")
-    with pytest.raises(ClientAuthenticationError, match=NO_ACCOUNTS):
+    with pytest.raises(CredentialUnavailableError, match=NO_ACCOUNTS):
         await SharedTokenCacheCredential(_cache=TokenCache(), tenant_id="not-cached").get_token("scope")
-    with pytest.raises(ClientAuthenticationError, match=NO_ACCOUNTS):
+    with pytest.raises(CredentialUnavailableError, match=NO_ACCOUNTS):
         credential = SharedTokenCacheCredential(_cache=TokenCache(), tenant_id="not-cached", username="not@cache")
         await credential.get_token("scope")
 
@@ -129,7 +131,7 @@ async def test_no_matching_account_for_username():
     account = get_account_event(username=upn, uid="uid", utid=tenant, refresh_token="refresh-token")
     cache = populated_cache(account)
 
-    with pytest.raises(ClientAuthenticationError) as ex:
+    with pytest.raises(CredentialUnavailableError) as ex:
         await SharedTokenCacheCredential(_cache=cache, username="not" + upn).get_token("scope")
 
     assert ex.value.message.startswith(NO_MATCHING_ACCOUNTS[: NO_MATCHING_ACCOUNTS.index("{")])
@@ -145,7 +147,7 @@ async def test_no_matching_account_for_tenant():
     account = get_account_event(username=upn, uid="uid", utid=tenant, refresh_token="refresh-token")
     cache = populated_cache(account)
 
-    with pytest.raises(ClientAuthenticationError) as ex:
+    with pytest.raises(CredentialUnavailableError) as ex:
         await SharedTokenCacheCredential(_cache=cache, tenant_id="not-" + tenant).get_token("scope")
 
     assert ex.value.message.startswith(NO_MATCHING_ACCOUNTS[: NO_MATCHING_ACCOUNTS.index("{")])
@@ -161,7 +163,7 @@ async def test_no_matching_account_for_tenant_and_username():
     account = get_account_event(username=upn, uid="uid", utid=tenant, refresh_token="refresh-token")
     cache = populated_cache(account)
 
-    with pytest.raises(ClientAuthenticationError) as ex:
+    with pytest.raises(CredentialUnavailableError) as ex:
         await SharedTokenCacheCredential(_cache=cache, tenant_id="not-" + tenant, username="not" + upn).get_token(
             "scope"
         )
@@ -187,13 +189,13 @@ async def test_no_matching_account_for_tenant_or_username():
     transport = Mock(side_effect=Exception())  # credential shouldn't use the network
 
     credential = SharedTokenCacheCredential(username=upn_a, tenant_id=tenant_b, _cache=cache, transport=transport)
-    with pytest.raises(ClientAuthenticationError) as ex:
+    with pytest.raises(CredentialUnavailableError) as ex:
         await credential.get_token("scope")
     assert ex.value.message.startswith(NO_MATCHING_ACCOUNTS[: NO_MATCHING_ACCOUNTS.index("{")])
     assert upn_a in ex.value.message and tenant_b in ex.value.message
 
     credential = SharedTokenCacheCredential(username=upn_b, tenant_id=tenant_a, _cache=cache, transport=transport)
-    with pytest.raises(ClientAuthenticationError) as ex:
+    with pytest.raises(CredentialUnavailableError) as ex:
         await credential.get_token("scope")
     assert ex.value.message.startswith(NO_MATCHING_ACCOUNTS[: NO_MATCHING_ACCOUNTS.index("{")])
     assert upn_b in ex.value.message and tenant_a in ex.value.message
@@ -290,11 +292,11 @@ async def test_no_refresh_token():
     transport = Mock(side_effect=Exception())  # credential shouldn't use the network
 
     credential = SharedTokenCacheCredential(_cache=cache, transport=transport)
-    with pytest.raises(ClientAuthenticationError, match=NO_ACCOUNTS):
+    with pytest.raises(CredentialUnavailableError, match=NO_ACCOUNTS):
         await credential.get_token("scope")
 
     credential = SharedTokenCacheCredential(_cache=cache, transport=transport, username="not@cache")
-    with pytest.raises(ClientAuthenticationError, match=NO_ACCOUNTS):
+    with pytest.raises(CredentialUnavailableError, match=NO_ACCOUNTS):
         await credential.get_token("scope")
 
 
@@ -311,7 +313,7 @@ async def test_two_accounts_no_username_or_tenant():
     # credential can't select an identity => it shouldn't use the network
     transport = Mock(side_effect=Exception())
 
-    # two users in the cache, no username specified -> ClientAuthenticationError
+    # two users in the cache, no username specified -> CredentialUnavailableError
     credential = SharedTokenCacheCredential(_cache=cache, transport=transport)
     with pytest.raises(ClientAuthenticationError, match=MULTIPLE_ACCOUNTS) as ex:
         await credential.get_token("scope")
@@ -404,7 +406,7 @@ async def test_same_username_different_tenants():
     # with no tenant specified the credential can't select an identity
     transport = Mock(side_effect=Exception())  # (so it shouldn't use the network)
     credential = SharedTokenCacheCredential(username=upn, _cache=cache, transport=transport)
-    with pytest.raises(ClientAuthenticationError) as ex:
+    with pytest.raises(CredentialUnavailableError) as ex:
         await credential.get_token("scope")
 
     assert ex.value.message.startswith(MULTIPLE_MATCHING_ACCOUNTS[: MULTIPLE_MATCHING_ACCOUNTS.index("{")])
@@ -448,7 +450,7 @@ async def test_same_tenant_different_usernames():
     # with no username specified the credential can't select an identity
     transport = Mock(side_effect=Exception())  # (so it shouldn't use the network)
     credential = SharedTokenCacheCredential(tenant_id=tenant_id, _cache=cache, transport=transport)
-    with pytest.raises(ClientAuthenticationError) as ex:
+    with pytest.raises(CredentialUnavailableError) as ex:
         await credential.get_token("scope")
 
     assert ex.value.message.startswith(MULTIPLE_MATCHING_ACCOUNTS[: MULTIPLE_MATCHING_ACCOUNTS.index("{")])