+++ sdk/identity/azure-identity/CHANGELOG.md
@@ -11,6 +11,18 @@ raising an error when none is passed. Although `get_token()` may sometimes
 have succeeded in prior versions, it couldn't do so consistently because its
 behavior was undefined, and dependened on the credential's type and internal
 state. ([#10243](https://github.com/Azure/azure-sdk-for-python/issues/10243))
+- `SharedTokenCacheCredential` raises `CredentialUnavailableError` when the
+cache is available but contains ambiguous or insufficient information. This
+causes `ChainedTokenCredential` to correctly try the next credential in the
+chain. ([#10631](https://github.com/Azure/azure-sdk-for-python/issues/10631))
+
+
+## 1.3.1 (2020-03-30)
+
+- `ManagedIdentityCredential` raises `CredentialUnavailableError` when no
+identity is configured for an IMDS endpoint. This causes
+`ChainedTokenCredential` to correctly try the next credential in the chain.
+([#10488](https://github.com/Azure/azure-sdk-for-python/issues/10488))
 
 
 ## 1.4.0b1 (2020-03-10)